+++ objects/Transaction.go
@@ -10,7 +10,7 @@ import (
 type Transaction struct {
 	From      PublicKey
 	To        PublicKey
-	Amount    int
+	Amount    uint64
 	ID        string
 	Signature string
 }
@@ -19,7 +19,7 @@ func (t Transaction) toString() string {
 	var buf bytes.Buffer
 	buf.WriteString(t.From.String())
 	buf.WriteString(t.To.String())
-	buf.WriteString(strconv.Itoa(t.Amount))
+	buf.WriteString(strconv.Itoa(int(t.Amount)))
 	buf.WriteString(t.ID)
 	return buf.String()
 }
@@ -33,7 +33,7 @@ func (t *Transaction) VerifyTransaction() bool {
 	return Verify(t.toString(), t.Signature, t.From)
 }
 
-func CreateTransaction(from PublicKey, to PublicKey, amount int, id string, sk SecretKey) Transaction {
+func CreateTransaction(from PublicKey, to PublicKey, amount uint64, id string, sk SecretKey) Transaction {
 	t := Transaction{from, to, amount, from.String()[4:14] + "-" + id + "-" + time.Now().String(), ""}
 	t.SignTransaction(sk)
 	return t
